+++ CHANGELOG.md
@@ -3,3 +3,11 @@
 All notable changes to this project will be documented in this file.
 
 This project adheres to [Semantic Versioning](http://semver.org/).
+
+## [0.0.3] - 2015-07-08
+### Changed
+- `addSource` to `connectSource`
+### Added
+- Support for media stream
+- Audio play toggling options
+- Custom colors option